General update

+++ TextureRenderShader.cpp
@@ -278,7 +278,7 @@ void CTextureRenderShader::ShutdownShader()
 void CTextureRenderShader::OutputShaderErrorMessage(ID3D10Blob* errorMessage, HWND hwnd, WCHAR* shaderFilename)
 {
 	char* compileErrors;
-	unsigned long bufferSize, i;
+	size_t bufferSize, i;
 	ofstream fout;
